+++ Changelog.md
@@ -3,13 +3,14 @@
 ## 0.6.6 ()
 
 - `HierarchicalMachine` now considers the initial state of `NestedState` instances/names passed to `initial`.
+- Bugfix: `HierarchicalMachine` used to ignore children when `NestedStates` were added to the machine.
 
 ## 0.6.5 (April, 2018)
 
 Release 0.6.5 is a minor release and contains a new feature and a bugfix:
 
 - Feature #287: Embedding `HierarchicalMachine` will now reuse the machine's `initial` state. Passing `initial: False` overrides this (thanks @mrjogo).
-- Bug #292: Models using `GraphMashine` were not picklable in the past due to `graph` property. Graphs for each model are now stored in `GraphMachine.model_graphs` (thanks @ansumanm)
+- Bugfix #292: Models using `GraphMashine` were not picklable in the past due to `graph` property. Graphs for each model are now stored in `GraphMachine.model_graphs` (thanks @ansumanm)
 
 ## 0.6.4 (January, 2018)